requirements.txt
pandas
pybliometrics
datetime
arxiv
openai
tqdm
python-dotenv
requests
habanero
beautifulsoup4
dvc 
dvc-s3 
mlflow